backend /requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.11.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
requests==2.31.0
pytesseract==0.3.10
pdf2image==1.17.0
pillow==10.2.0
sqlalchemy==2.0.27
mysqlclient==2.2.4
python-magic==0.4.27
aiofiles==23.2.1
pydantic-settings>=2.0
pdfplumber
paddlepaddle
paddleocr
docling
deepsearch-toolkit
duckduckgo-search
beautifulsoup4